Refactor:
In order for this to work, we'd need to extract the cookie handling from
the session state, and right now MemoryCache works just fine if we're
also authenticated.

However, with the refactor to save state in cookies, we'll be able to do
this refactor, then continue with this refactor in full.